backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osprey-backend"
version = "2.0.0"
description = "Osprey Immigration Platform - AI-powered visa application processing"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Osprey Team", email = "dev@osprey.ai"}
]
keywords = ["immigration", "visa", "ai", "fastapi", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    # Core Framework & Web Server
    "fastapi==0.128.0",
    "starlette>=0.50.0,<0.51.0",
    "uvicorn==0.40.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    
    # Database & Storage
    "motor==3.7.1",
    "pymongo==4.16.0",
    "dnspython==2.8.0",
    
    # AI & Machine Learning
    "openai>=2.15.0",
    "portkey-ai>=2.1.0",
    "google-genai>=1.39.1",
    "google-generativeai>=0.8.5",
    "google-ai-generativelanguage>=0.6.15",
    "litellm>=1.77.4",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    
    # Google Cloud Services
    "google-cloud-documentai>=3.6.0",
    "google-cloud-vision>=3.10.2",
    "google-cloud-speech>=2.26.0",
    "google-cloud-texttospeech>=2.24.0",
    "google-api-core>=2.25.1",
    "google-api-python-client==2.187.0",
    "googleapis-common-protos==1.72.0",
    "google-auth>=2.40.3,<3.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "grpcio>=1.75.1",
    "grpcio-status>=1.71.2",
    
    # Security & Authentication
    "PyJWT==2.10.1",
    "bcrypt==5.0.0",
    "python-jose==3.5.0",
    "cryptography==46.0.3",
    "passlib==1.7.4",
    "ecdsa==0.19.1",
    
    # Payment Processing
    "stripe==12.5.1",
    
    # Email
    "resend==2.19.0",
    "email-validator==2.3.0",
    
    # Document Processing & PDFs
    "PyMuPDF==1.26.7",
    "PyPDF2==3.0.1",
    "pypdf==6.6.0",
    "reportlab==4.4.7",
    "fpdf2==2.8.5",
    "pdfplumber",
    "python-docx>=0.8.11",
    
    # HTTP & Networking
    "requests==2.32.5",
    "aiohttp==3.13.3",
    "httpx==0.28.1",
    "httpcore==1.0.9",
    "urllib3==2.6.3",
    "requests-oauthlib==2.0.0",
    "oauthlib==3.3.1",
    "aiohappyeyeballs==2.6.1",
    "aiosignal==1.4.0",
    "multidict==6.7.0",
    "yarl==1.22.0",
    "frozenlist==1.8.0",
    "propcache==0.4.1",
    "websockets>=15.0.1,<15.1.0",
    "h11==0.16.0",
    "httplib2==0.31.1",
    
    # Data Processing & Analysis
    "numpy==2.4.1",
    "pandas==2.3.3",
    
    # Image Processing
    "Pillow==12.1.0",
    "fonttools==4.61.1",
    
    # Templating & Markup
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "beautifulsoup4==4.14.3",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    
    # Data Validation & Serialization
    "PyYAML==6.0.3",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "referencing==0.37.0",
    "protobuf>=5.26.1,<6.0.0",
    "proto-plus==1.27.0",
    "defusedxml==0.7.1",
    
    # Background Jobs & Scheduling
    "APScheduler==3.11.2",
    "tenacity==9.1.2",
    
    # Monitoring & Observability
    "sentry-sdk[fastapi]==2.20.0",
    "rich==14.2.0",
    "Pygments==2.19.2",
    
    # Utilities & Helpers
    "click==8.3.1",
    "typer==0.21.1",
    "python-dotenv==1.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "tzdata==2025.3",
    "regex==2025.11.3",
    "six==1.17.0",
    
    # AWS SDK
    "boto3==1.42.26",
    "botocore==1.42.26",
    "s3transfer==0.16.0",
    "jmespath==1.0.1",
    
    # File & Path Utilities
    "watchfiles==1.1.1",
    "filelock==3.20.3",
    "pathspec==1.0.3",
    "platformdirs==4.5.1",
    "fsspec==2026.1.0",
    
    # Miscellaneous
    "python-multipart==0.0.21",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "idna==3.11",
    "anyio==4.12.1",
    "sniffio==1.3.1",
    "annotated-types==0.7.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "importlib-metadata==8.7.1",
    "zipp==3.23.0",
    "attrs==25.4.0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "iniconfig==2.3.0",
    "pyparsing==3.3.1",
    "shellingham==1.5.4",
    "tqdm==4.67.1",
    "jiter==0.12.0",
    "cffi==2.0.0",
    "pycparser==2.23",
    "distro==1.9.0",
    "cachetools==6.2.4",
    "uritemplate==4.2.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "rsa==4.9.1",
    
    # Hugging Face
    "huggingface-hub==1.3.1",
    "hf-xet==1.2.0",
    
    # Specialized Libraries
    "onfido-python==5.5.0",
    "jq==1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "black==25.12.0",
    "flake8==7.3.0",
    "pycodestyle==2.14.0",
    "pyflakes==3.4.0",
    "mccabe==0.7.0",
    "isort==7.0.0",
    "mypy==1.19.1",
    "mypy-extensions==1.1.0",
]

[project.urls]
Homepage = "https://osprey.ai"
Documentation = "https://docs.osprey.ai"
Repository = "https://github.com/osprey/osprey-core"
Issues = "https://github.com/osprey/osprey-core/issues"

[tool.setuptools]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/venv/*", "*/.venv/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "pymongo.*",
    "google.*",
    "portkey_ai.*",
    "litellm.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "9.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["backend"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
    "*.egg-info",
]
max-complexity = 10
